Change the data structure of cachedAgents. (#36)
Specifically, cachedAgents[this.username] now stores a data structure
consisting of a v10 user agent and/or a v11 user agent, plus shared
state. The shared state is used predominantly for session details, as
it ensures that a refresh against a v10 user will also affect the v11
user.

This also fixes a nasty bug where using Agent.on() multiple times would
clone repeatedly, using lots and lots of memory.